# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Weather Radar API",
    "version": "5.0.0"
  },
  "paths": {
...
    "/api/radar/atlas/{timestamp}/{z}/{x}/{y}.png": {
      "get": {
        "summary": "Radar Atlas Tile",
        "description": "Serve a 256\u00d72048 grayscale PNG atlas tile with 8 tilt bands.",
        "operationId": "radar_atlas_tile_api_radar_atlas__timestamp___z___x___y__png_get",
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timestamp"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/radar/atlas-raw/{timestamp}/{z}/{x}/{y}.bin": {
      "get": {
        "summary": "Radar Atlas Tile Raw",
        "description": "Serve a 256\u00d72048 single-channel uint8 atlas tile as raw bytes.\n\nIdentical payload to the ``.png`` endpoint but undecoded, so React Native\n(which lacks ``createImageBitmap``/canvas) can upload it straight to an R8\n``DataTexture`` without a client-side PNG decode.",
        "operationId": "radar_atlas_tile_raw_api_radar_atlas_raw__timestamp___z___x___y__bin_get",
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timestamp"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/radar/motion/{timestamp}.png": {
      "get": {
        "summary": "Radar Motion Tile",
        "description": "Serve the motion vector field for *timestamp* \u2192 next frame as RGB PNG.",
        "operationId": "radar_motion_tile_api_radar_motion__timestamp__png_get",
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timestamp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/radar/motion-raw/{timestamp}.bin": {
      "get": {
        "summary": "Radar Motion Tile Raw",
        "description": "Serve the motion field as raw RGB bytes (for native DataTexture upload).\n\nDecodes the stored motion PNG server-side so React Native doesn't need a PNG\ndecoder. Width/height are returned in the X-Width / X-Height headers.",
        "operationId": "radar_motion_tile_raw_api_radar_motion_raw__timestamp__bin_get",
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timestamp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}